+++ packages/ethereum-contracts/CHANGELOG.md
@@ -9,9 +9,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - `batchCall` supports new `send` batch operation
 - Added `downgradeTo` function in `SuperToken.sol`
 
+### Breaking
+
+- `scripts/deploy-test-framework.js` no longer exists, `deploy-test-framework.js` has been moved to `dev-scripts`
+  - Migration: Use `dev-scripts/deploy-test-framework.js` instead
+
 ### Changed
 
 - `upgradeTo` logic changed to revert if `data` is not empty and `to` is a contract and is not a registered ERC777 recipient
+- `MAX_APP_CALLBACK_LEVEL` is public again
 
 ### [v1.4.3] - 2022-10-27
 ### Added